LiveTranslate/requirements.txt
numpy==1.26.4
pyaudio
vosk
keyboard
schedule
google-generativeai
python-dotenv
pyttsx3
argostranslate
simpleaudio
pyautogui
